feat(wip): error handling is bad(
