+++ server/Cargo.toml
@@ -18,11 +18,12 @@ tonic = "0.13.1"
 tracing = "0.1.41"
 tracing-subscriber = { version = "0.3.19", features = ["env-filter"] }
 
+uuid = { version = "1.17.0", features = ["v7"] }
+
 evops-models = { path = "../server-ext/client-ext/crates/evops-models/" }
 evops-db = { path = "../server-ext/crates/evops-db/" }
 url = "2.5.4"
 bon = "3.6.4"
-
 [build-dependencies]
 eyre = "0.6.12"
 tonic-build = "0.13.1"
