+++ server/src/main.rs
@@ -2,14 +2,10 @@ use tonic::transport::Server;
 
 use std::net::{Ipv4Addr, SocketAddr};
 
-use crate::service::MlService;
+use crate::pb::AppState;
 use crate::shutdown::signal;
 
-use std::sync::Arc;
-
-use bon::bon;
-use eyre::Context as _;
-use url::Url;
+use crate::service::Service;
 
 use const_format::formatcp;
 use tracing::{debug, info};
@@ -34,7 +30,17 @@ async fn main() -> eyre::Result<()> {
         debug!(".env not found, using environment variables");
     }
     let config = self::config::from_env()?;
-    let service = self::MlService::default();
+
+    let state = {
+        let builder = AppState::builder().database_url(&config.database_url);
+
+        match config.auto_tags_treshhold {
+            Some(val) => builder.auto_tags_treshhold(val).build().await,
+            None => builder.build().await,
+        }?
+    };
+
+    let service = self::Service { state };
 
     let addr = SocketAddr::new(Ipv4Addr::UNSPECIFIED.into(), config.port);
     info!("listening on port {}", config.port);
@@ -69,43 +75,4 @@ fn init_logging() {
         .with(env_filter)
         .with(fmt_layer)
         .init();
-}
-
-struct State {
-    db: tokio::sync::Mutex<evops_db::Database>,
-}
-
-#[derive(Clone)]
-pub struct AppState {
-    // Here, the `State` struct, defined above, only gets wrapped in an `Arc` once.
-    shared_state: Arc<self::State>,
-}
-
-#[bon]
-impl AppState {
-    #[builder]
-    pub async fn new(database_url: &Url) -> eyre::Result<Self> {
-        let db = {
-            evops_db::Database::establish_connection(database_url)
-                .await
-                .wrap_err("error connecting to db")?
-        };
-        Ok(Self {
-            shared_state: {
-                Arc::new(self::State {
-                    db: tokio::sync::Mutex::new(db),
-                })
-            },
-        })
-    }
-
-    /// Does the same thing as `self.clone()`,
-    /// but the method name explicitly tells that the new object
-    /// will point to the same memory location.
-    #[must_use]
-    pub fn arc_clone(&self) -> Self {
-        Self {
-            shared_state: Arc::clone(&self.shared_state),
-        }
-    }
 }
