+++ server/src/service/auto_tags.rs
@@ -2,6 +2,8 @@ use std::collections::HashMap;
 
 use evops_models::{ApiResult, Tag};
 
+use uuid::Uuid;
+
 impl crate::AppState {
     async fn call_auto_tags(
         &self,
@@ -11,7 +13,7 @@ impl crate::AppState {
         Ok(Vec::new())
     }
 
-    pub async fn get_tags(&self, description: String) -> ApiResult<Vec<String>> {
+    pub async fn get_tags_et(&self, description: String) -> ApiResult<Vec<Uuid>> {
         let tags_list = {
             let mut db = self.shared_state.db.lock().await;
             db.list_tags(None, None).await
@@ -23,6 +25,7 @@ impl crate::AppState {
         let correct_tags = self
             .call_auto_tags(description, tags_map.keys().clone().collect())
             .await?;
+        let result = correct_tags.into_iter().map(|tag| tag);
         Ok(Vec::new())
     }
 }
